Tests/LibWeb/Text/expected/UIEvents/KeyEvent-cancelled.txt

keydown A
cancel keydown
You can't change me!
keydown B
keypress B
cancel keypress
You can't change me!
keydown C
keypress C
C
